backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
aiosqlite==0.19.0
python-multipart==0.0.9
alembic==1.13.1
sse-starlette==2.0.0
websockets==12.0
